offshore_detector/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pandas==2.1.4
openpyxl==3.1.2
xlsxwriter==3.1.9
python-Levenshtein==0.25.0
openai==1.12.0
pydantic==2.6.0
tenacity==8.2.3
python-dotenv==1.0.0
aiofiles==23.2.1